Reload accessibility service infos for reisntalled packages.
When something that affects the state of accessibility in the sysytem
changes, we run a reolve method that reloads all relevant information and
if it changed we call a method that makes everyting right. One of the
interesting properties we read is the isntalled accessibliity services.

We are using equals to figure out whether these services have changed
but this is not correct since AccessibilityServiceInfo does not use all
internal members for equals and using all memthis is not reasible since
some of these internal members do not support equals propertly, for
example ResolveInfo.

Therefore, when a package is reinstalled we remove all installed services
from the list of ones we know about which forces them to be reloaded,
thus capturing the current state of a reinstalled package.

bug:8621960

Change-Id: Ie1ef4bf1036d8d6e033cd9528ea2292ce24e5320

Former-commit-id: 0ecfa017decd230051d5e42d9f49129afd58838b