Provide SharedPreferences coherence guarantees for BackupAgent
SharedPreferences uses deferred writes internally, and the public
API doesn't allow apps to explicitly synchronize with this, so the
backup/restore implementation needs to take a little care to make
sure that the app process isn't killed before the deferred writes
land on disk.  This parallels the coherence guarantees around
SharedPreference that the Activity and Service lifecycles provide.

Bug 8659368

Change-Id: I853e54f9fb0d2d260dbe6e40d640959f998092df

Former-commit-id: f85f5b2125461ea664cf67a16d4608a5a9bf2f98