Fix bug where owner info wasn't being updated properly.
There was a bug where older platforms only stored the 'enabled'
value if the checkbox in Settings was toggled at least once. This
caused the checkbox to be checked off when upgrading the device.

The code now mimicks the old behavior by detecting a non-empty
info string and setting 'enabled' appropriately.

Fixes bug 8659048

Change-Id: If2431d54f168a0aab286628165dd3f6b4a0d5cf8

Former-commit-id: 2d8ecf9df0ba4f503a09e87f91d12433f29faf52