Perform idle maintenance when the device is charging.
Added the precondition that the device should be charging to the
existing ones of the device not being used for awhile and the
battery level being high enough. Note that even if the device is
charging, we have to check the battery level since the user can
unplug it at any time.

bug:8688454

Change-Id: I709b1620571301743dc3504a6a625e2018951bfa

Former-commit-id: f23b64df94bbff9ac8c87832dea9c1bc0ae9a950