Revert change 116a7025605e63996b4ffb2fbed92501403f72e5 [formerly f449c0c381239dc0f78b0f01ecfb3889e056a597].
Previously, onHoverEvent() would return true if a view was hoverable
and consume the event. After the change, it would return the result of
dispatchGenericMotionEventInternal(). As a result, touch exploration
caused multiple hover events to be sent from every view under a given
touch point. This change reverts to the original behavior and fixes
touch exploration.

BUG: 8723842
Change-Id: I0c7362f19c51bf21ed842711a03b7f02613958d2

Former-commit-id: e7a4e636ac0923d72505efd6df8cd130c83fe97c