Fix network reload when config is restored
With scan mode opted in, supplicant connection is not shut
down even when wifi is turned off. This is a problem since
networks need to be reloaded when wifi is turned off and turned on
and this currently happens only on a supplicant connection.

Fix to handle this for scan mode state.

Bug: 8714796
Change-Id: I7d66c39fb51018fb52e783341222cea993993893

Former-commit-id: 428b063a2e1f5c9bb8caf560e770ea5f9d5fde02