Remove Auto-rotate screen setting if a required feature is missing.
It is possible that a device running Android does not support rotation,
for example a Google TV box. The change adds a check to the rotation
policy to take this into account.

The SystemUI was not respecting the fact that the user has turned off
screen rotation, i.e. locked the rotation, from the accessibility
settings. In this case we should not show the affordance to toggle
screen rotation in the quick settings to minimize the risk of a blind
user accidentally turning screen rotation on. For a blind person
screen rotation adds 2X complexity since he/she should learn two
different layouts for the app.

bug:8051556

Change-Id: If1b3c092476932f91b0345fb9bfbbf6ed33d0df9

Former-commit-id: 79578b29bf4ba1d210586b9d7bb832eddf0960b7