Eliminate flash when going from home to recents
When the wallpaper animation is coupled to the recents animation it
will go from fully transparent to visible along with it. In master
this is causing noticable jank if we are going from the home screen
where it was previously visible. Detaching the wallpaper animation
from the recents animation fixes it.

Change-Id: If1d1a1478530610017d954983b00ca6101cbaa8b

Former-commit-id: c20054a16691ddba0a885351c63e6bf7dcbd2faa