Do not report the fake UI automation service to clients.
For UI test automation purposes we register a fake accessibility service
and suspend all other services. When the UI automation serivce is unregistered
we restore the suspended ones. Since the UI automation serivce is fake and
incomplete, for example it has not resolve info, it should not be reported
to clients as being installed or enabled.

bug:8871034

Change-Id: I66792cd028159c1652d3c8a2982164821282ab24

Former-commit-id: 319d55bd15721982db7563711c876bb000490b7d