Fix out-of-screen item can take focus issue
In case of a ListView has a item which takes more than 33% of its height
(result from getMaxScrollAmount), scroll does not move enough to show
all the item. In this case, consecutive arrow scroll moves selection
everytime while scroll may not move that much.

This fixes the issue by checking the visibility of current selection in
advance.

Bug: 8831751
Change-Id: Ic747bd5513c6734aaf1a1d08e497c1e3ef835004

Former-commit-id: 49fc4dcd4c0d88b74084d6b2a2d843ad17b8a1eb