Do not block notifications or toasts for SYSTEM_UID or PHONE_UID.
The specific bug is this: SIM PIN unlock attempt toasts are
sent from com.android.settings/.IccLockSettings which runs
as the phone process; NoMan wasn't having any of that and
was blocking the toasts.

With this change we treat SYSTEM_UID and PHONE_UID the same
for all security checks, and furthermore we guarantee that
all notifications and toasts from those UIDs will be
permitted.

Bug: 9098802
Change-Id: Idc788527aa2cb38e015fe92773766a514167999e

Former-commit-id: 9698fa4c38bcbda13371163012b4af400e30f7da