Propagate pivot values of 0 to native layer
Logic in pivotXY setters noops when the new value equals
the previous value. However, the initial value is "0" even
though we actually use a value of the view's midpoint by
default. If an app sets a new value of 0, we don't send it
down to the native layer because it's the same as the initial
value, even though we're actually using a midpoint value instead.
This causes a conflict between the matrix used for invalidations
(which use the actual values the app set) and the matrix used
for rendering (which uses the default midpoint values).

The fix is to make sure we send down the initial value, even when it
equals the default value, by checking to see whether this is the
first time we're setting the pivot.

Issue #9337635 Clipping and bad rendering of view corners when y pivot is set

Change-Id: I4aa20c4a3c9a866ca17df3e067232b832d0ef504

Former-commit-id: 183e2a351daa8071be69eacb514b9405193a7a35