TelephonyManager.listen can have notifyNow always true.
Since telepony.registry is a real system service notifyNow
parameter doesn't need to be conditional as telephony.registery
will never go away.

This is different from most of the other TelephonyManager
methods which are used to invoke methods on the phone service
which implements ITelephony and is implemented by
PhoneInterfaceManager in the phone application. Since the
phone app is not a system service it can and does go away when
it crashes.

Bug: 9393863
Change-Id: I1a8afc12b0e139e72f05820e49f3d996aec2b52a

Former-commit-id: 42b04e4cf1935921fe90f9a8eab3327d723435b0