move scoring, sound, vibration, and lights into a handler
As a prelude to introducing fancier scoring functions, this CL
moves scoring and everything that depends on it into a handler.
This allows the NotificationManagerService to report back to the
calling process right away about whether or not the notification
is valid; not blocking while score is computed, and non-statusbar
notification modalities are activated.

Change-Id: Iaf77bc89544bb7bcb7adeda60885d74c3170541c

Former-commit-id: 9b05c61a87972dce6982e5b016bbcd0e0dff74c3