Always show overflow buttons on action bars
The menu button is not pining, it's passed on! This button is no more!
It has ceased to be! It's expired and gone to meet its maker! It's a
stiff! Bereft of life, it rests in peace! If you hadn't nailed it to a
flagship device it'd be pushing up the daisies!

THIS IS AN EX-BUTTON!!

Change-Id: Id523ef8efa7d0fa7120f8430222ac892fcafefda

Former-commit-id: ea04f3cfc6e245fb415fd352ed0048cd940a46fe