Allowing the root user to use the screen introspection APIs.
We were allowing the system and the shell user to access use the
screen introspection APIs but the root user was not able to to so.
This change enableS the root user to also use these APIs. Note that
we usually allow the root user to access privileged functionality
similarly to the shell user of the system.

bug:8877685

Change-Id: Ie4008339e864b835bd3a2d5e06b042e4431c5270

Former-commit-id: d08d7ee4e9ce8c388e5dba392e2a6b156180dabc