Add WebView.evaluateJavaScript(String, ValueCallback) API.
This API is intended to replace the "WebView.loadUrl("javascript:...")
pattern that is the current mechanism for executing JavaScript in the
context of the current page displayed in the WebView.

The new API is more convenient - it doesn't trigger the normal URL
loading path - and so does not have side effects such as hiding
the keyboard - and allows the caller to specify a callback that will
be invoked once the script execution is complete.

BUG=9814043

Change-Id: I3f27e8ff5371077d9265430090d61381a3a86e76

Former-commit-id: a5cdd51cb7dc3acd608b67186bdef1d2a12bdbe7