Changed name of recent hasLayout() method to isLaidOut()
This new method on view reflects whether the view has been laid out
at least once since it was attached. hasLayout() seems too vague for that
meaning; every View that has a parent has a layout (since we use container,
parent, and layout interchangeably). The new version of the method
is closer to the actual meaning.

Change-Id: I519745739b6a6317faeb077aa61f994025cf81f3

Former-commit-id: 7a46dde1ae56a85fcb5cdac91173424b6355bf3c