Track latency of captive portal checks
When captive portal check occurs, track its latency, whether or not
we received a response, and whether or not the response was a captive
portal.  Pair with information identifying the access point / base
station, and broadcast it (with a system|signature-protected
permission).

Broadcast only occurs if user has consented to
Settings.Global.WIFI_SCAN_ALWAYS_AVAILABLE.

Change-Id: I6fd59954a7ee2cc7acedf064a1465882653b2173

Former-commit-id: 1ed51627d94b140e8893b8e1dcde9bf37e2adca7