Add a new process state for "service restarting".
This is a process that is not actually running, but would be
if we had enough RAM.

Also rework how service stats are stored to used the nice
compact structure we are already using for processes.

And fix a bug where when we committed the current state data,
we would mistakenly write it under the name of the *next* state
that is now starting.  Ouch.

Change-Id: I9f2b84b8b3f305301fd48dcd9d4e6c232abe8ef9

Former-commit-id: 8eea0fcd3b4a9c58ba70ea80ed0332d37ac43da5