Add extra layout pass after draw finished.
Once content has been drawn another pass through layout is required
to set mHasContent in the LogicalDisplay. Previously this pass was
occuring because of a delayed animation step. When timing of that
step changed that pass occurred before the draw completed. This is
why Presentations were immediately displayed in jb-mr1 and not
jb-mr2.

Fixes bug 10154780.

Change-Id: I0075c5a73d5cdf972e73fdd59c1fde46df64e245

Former-commit-id: 8efb0a4d811fc18ac8ef32f7d597aa6fafb3acec