Allow bound processes to go PROCESS_STATE_TOP.
For process state, if a top process is making use of another
process, we should probably count the second process as top
as well (instead of IMPORTANT_FOREGROUND).  Specially, when
chrome binds to render processes we want those render
processes to be TOP.  Otherwise, they end up in the
important foreground state and it looks like they are running
for some other reason in the background.

Change-Id: Id115dbb65dc2b403ffa4bbe4e7837564eb3b9cb5

Former-commit-id: d9545a913da2ac705dde33a196576d37fe760796