Display signal strength on data-only SIMs.
On SIM cards that support data but not voice, we currently
display a null signal strength icon, which is confusing because
data is actually working.

Fix this by displaying signal bars as well as showing the
"emergency calls only" or "no service" text which indicates that
voice service is not available.

Bug: 3339315
Change-Id: I7d888721e8dc5e22fefd1b8fa85ba046d46a8fba

Former-commit-id: 2ee168ea02170a542f848c5c55ef969e7a237545