Not create context menu when mouse right click outside list/grid items
Even if mouse right click outside of list/grid items, a context menu
is made. But there is no item of the list, so a context menu info is
null. Then, NullPointerException happens at onCreateContextMenu()
using a context menu info.
For fixing the problem, check mTouchMoude and mMotionPosition to
distingusih where is clicked, on a list item or not, before
performButtonActionOnTouchDown() which makes a context menu.

Manually merged from AOSP contribution.

Change-Id: I778be73df36abbfd4ba9534d8e7a1ece6844ed40
Author: henrik baard <henrik.baard@sonymobile.com>

Former-commit-id: b339cc5bcbafbc5b9cf184c6f87b4eb466386ea6