Make "low ram device" a readonly system property
This lets components outside the Zygote / Android resource regime
know whether this is considered a "low ram" device.

Set the property [ro.config.low_ram] to "true" if the device is to be
considered low-ram.  Any other property string (or its absence) will
be interpreted as meaning this is a normal larger-ram device.  The
"true" string must be all lower case.

Bug 10131789

Change-Id: I15332d169b4aa4eaa0a6c1952787af0b03d23a6b

Former-commit-id: 9fc5baed6f1e6f4ce089a6011170657183245d60