Camera2: Updated Rational for denom=0.
Handle NaN and INF using the same semantics as Java floats. No longer
throw an exception if a rational is constructed with a zero-denominator.

This commit is required for the ITS pdk app, since without it, the app
fails when it tries to deserialize rational values that are
(incorrectly) set to 0/0.

Bug: 10337323

Change-Id: I890ec2092f2945abc0f531fc46f9a7f5c71d2dbf

Former-commit-id: 23c8809568b68430f875df759108c58016c44c16