When waiting for activity to finish don't reset
In cases where the client is waiting for an activity to launch
(startActivityMayWait()) it is a bad idea to clear
ActivityRecord.displayStartTime when going into the pause state. If
displayStartTime is cleared before the activity is displayed,
the client will never be released.

This fix keeps pause from clearing displayStartTime if any client
is waiting for the activity to be displayed.

Fixes bug 10095558. But not a permanent fix, startActivityMayWait()
should not be called by any production code.

Change-Id: I7cbdcb04256f4a26233867c52aedd3bc4151adc3

Former-commit-id: 5c494547af1e4558aea5994b60afaadefef971e1