Update documentation for speech APIs.
Makes it clear that the intent and service APIs are likely to stream
audio over the network and notes that impact this has on bandwidth and
battery life.
Also clarifies that the service API should not be used for continuous
recognition as a result. The intent API only runs when the activity is
in the foreground so I didn't mention continuous recognition for it.

Bug: 10674392
Change-Id: Ib5e0c16b3124f4b7d892ef5bd4167f23662d2ca3
(cherry picked from commit b1c48947075e1526ef398499db6e2342ac45d019 [formerly 2921cee3048f7e64ba6645d50a1c1705ef9658f8])
Former-commit-id: 76b02e4e14f5326a0cf507ae27525be575ab42ba