Check RelativeLayout's access of its own LayoutParams during measurement
RelativeLayout's use of its own LayoutParams is really old behavior
and something it probably never should have done in the first
place. However at this point it's enshrined compatibility behavior as
fixing this would change the end results of measurement in some cases,
and we've found that apps really love doing strange things with
RelativeLayout that hit all of those edge cases.

The existing behavior ignores this code path if set to MATCH_PARENT or
WRAP_CONTENT, so we'll simply treat missing LayoutParams as the same
case.

Bug 10733375

Change-Id: I8b40b8557d19695ebcd5329aefc1603e2bf4234e

Former-commit-id: 758d5a4563a71d4f3f7efbf7e450e8e537abc185